moved aws inside the server folder Co-authored-by: aelizj <anne1365@outlook.com> Co-authored-by: Pauline3414159 <fagie37@hotmail.com> Co-authored-by: olgashi <oshiryaeva@protonmail.com> Co-authored-by: donald-p-redding <donald.p.redding@gmail.com>
