+++ server/aws/.gitignore
@@ -1,0 +1,4 @@
+node_modules
+.env
+libs/_unused
+libs/sendObjectToLogstashHTTPPlugin.mjs
