+++ server/aws/constants/general.mjs
