+++ server/aws/constants/lambdaAWSPolicies.mjs
