+++ server/aws/constants/lambdaDeploymentPackage.mjs
