+++ server/aws/lambdaDeploymentPackage/nodejs/index.js
