+++ server/aws/lambdaDeploymentPackage/nodejs/package.json
