+++ server/aws/libs/iam/attachMultipleRolePolicies.mjs
