+++ server/aws/libs/lambda/deleteLambda.mjs
