+++ server/aws/libs/s3/deleteObjectsInS3Bucket.mjs
