+++ server/aws/libs/s3/downloadObjectFromBucket.mjs
