+++ server/aws/libs/s3/getObjectListFromBucket.mjs
