+++ server/aws/libs/s3/uploadObjectToBucket.mjs
