+++ server/aws/package.json
